Isolate "pause on gameplay pause" logic to its own sub-class
